# backend/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Resume Customizer API",
    "description": "API for customizing resumes based on job descriptions using AI",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/ghodeaniket/Customize-Resume-with-AI"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.your-domain.com/api/v1",
      "description": "Production API server"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development API server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "API key authentication. Format: Bearer YOUR_API_KEY"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["status", "message"],
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Error message description"
          },
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "details": {
            "type": "object",
            "example": {
              "field": "Invalid field value"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2023-05-15T10:30:00Z"
          }
        }
      },
      "JobStatus": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "processing", "completed", "failed"],
            "example": "completed"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-05-15T10:30:00Z"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-05-15T10:35:00Z"
          },
          "result": {
            "type": "string",
            "example": "Customized resume content"
          },
          "error": {
            "type": "string",
            "example": "Error message if job failed"
          }
        }
      },
      "ResumeCustomizationRequest": {
        "type": "object",
        "required": ["resumeContent", "jobDescription"],
        "properties": {
          "resumeContent": {
            "type": "string",
            "description": "The content of the resume, either as plain text or base64-encoded file data",
            "example": "John Doe\\nSoftware Engineer\\njohn.doe@example.com | (123) 456-7890\\n\\nEXPERIENCE\\n..."
          },
          "jobDescription": {
            "type": "string",
            "description": "The job description text or a URL to the job posting",
            "example": "We are looking for a Senior Software Engineer..."
          },
          "resumeFormat": {
            "type": "string",
            "enum": ["text", "pdf", "docx", "html", "json"],
            "default": "text",
            "description": "Format of the resume content"
          },
          "isJobDescriptionUrl": {
            "type": "boolean",
            "default": false,
            "description": "Set to true if jobDescription is a URL"
          },
          "profilerModel": {
            "type": "string",
            "description": "OpenRouter model ID for the profile generation step",
            "example": "anthropic/claude-3-opus"
          },
          "researcherModel": {
            "type": "string",
            "description": "OpenRouter model ID for the job description analysis step",
            "example": "anthropic/claude-3-opus"
          },
          "strategistModel": {
            "type": "string",
            "description": "OpenRouter model ID for the resume customization step",
            "example": "anthropic/claude-3-opus"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Too many requests",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
    "/resume/customize": {
      "post": {
        "summary": "Submit a resume for customization",
        "description": "Submits a resume for customization based on a job description",
        "operationId": "customizeResume",
        "requestBody": {
          "description": "Resume and job description details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeCustomizationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resume customization job submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Resume customization job submitted successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobId": {
                          "type": "string",
                          "format": "uuid",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/resume/status/{jobId}": {
      "get": {
        "summary": "Get job status",
        "description": "Get the status of a resume customization job",
        "operationId": "getJobStatus",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "ID of the job to check",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/JobStatus"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/resume/history": {
      "get": {
        "summary": "Get job history",
        "description": "Get a list of recent resume customization jobs for the current user",
        "operationId": "getJobHistory",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobs": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/JobStatus"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "total": {
                              "type": "integer",
                              "example": 25
                            },
                            "page": {
                              "type": "integer",
                              "example": 1
                            },
                            "limit": {
                              "type": "integer",
                              "example": 10
                            },
                            "totalPages": {
                              "type": "integer",
                              "example": 3
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/formatted-resume/customize": {
      "post": {
        "summary": "Submit a resume for formatted customization",
        "description": "Submits a resume for customization with formatting preservation",
        "operationId": "customizeFormattedResume",
        "requestBody": {
          "description": "Resume and job description details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeCustomizationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Formatted resume customization job submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Formatted resume customization job submitted successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobId": {
                          "type": "string",
                          "format": "uuid",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/formatted-resume/status/{jobId}": {
      "get": {
        "summary": "Get formatted job status",
        "description": "Get the status of a formatted resume customization job",
        "operationId": "getFormattedJobStatus",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "ID of the job to check",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Formatted job status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/JobStatus"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/formatted-resume/pdf/{jobId}": {
      "get": {
        "summary": "Get formatted PDF result",
        "description": "Get the formatted PDF result of a completed job",
        "operationId": "getFormattedPdfResult",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "description": "ID of the job",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF document",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  }
}